Can't figure out GitHub actions, so we're going dumb
..and we're committing the entire /public/ directory.
Ideally, we would just commit the files needed to build with
(i.e. template and documents) and use GitHub Actions to
build with Hugo/Go and push to the right branch.
However, I can NOT figure out how to do this for my LIFE.

So instead, we're going old school.

By building locally, pushing everything, and telling GitHub to
read from the directory /public/, we're leaving zero ambiguity
for deployment, but we're polluting the heck out of our git history.

Oh well.